Separate baseof file for AMP.
Initially I wanted to have just one site variant: AMP. But maybe I can
have two, HTML and AMP, with AMP sitting in a subdirectory.

I'm still figuring this out. So far it seems like the way it should
work: HTML files sit in the root directory and define the list of links.
The HTML files define the list of links, and provide a link rel=amphtml
pointing to the corresponding page in the amp/ subdirectory.

I'm also experimenting with blocks to only define the difference between
the two formats, without repeating parts of templates.

The remaining bit is how to support the amp-img tag from the figure
shortcode. I thought it would work automatically thanks to
https://github.com/gohugoio/hugo/issues/3220 but I tried it and it
doesn't seem to work automatically.